API: Move NLM functional to contrib
