+++ odl/contrib/solvers/README.md
@@ -1,0 +1,4 @@
+# odl.contrib.solvers
+
+- Contributed additions to the [odl.solvers](../../solvers) package.
+- Contains application specific functionals and solvers
