+++ odl/contrib/solvers/functional/__init__.py
@@ -10,15 +10,5 @@ from __future__ import absolute_import
 
 __all__ = ()
 
-from .functional import *
-__all__ += functional.__all__
-
-from .default_functionals import *
-__all__ += default_functionals.__all__
-
-from .example_funcs import *
-__all__ += example_funcs.__all__
-
-from .derivatives import *
-__all__ += derivatives.__all__
-
+from .nonlocalmeans_functionals import *
+__all__ += nonlocalmeans_functionals.__all__
