+++ odl/contrib/solvers/functional/nonlocalmeans_functionals.py
